requirements.txt
fastapi
uvicorn
jinja2
requests
feedparser
beautifulsoup4
deep-translator
pillow
httpx
pytest
pytest-asyncio
